+++ lib/src/postback_native.dart
@@ -83,16 +83,7 @@ class PostbackNative {
     return DeviceInfo.fromJson(result ?? const <dynamic, dynamic>{});
   }
 
-  static Future<String?> getWebViewUserAgent() {
-    return _channel.invokeMethod<String>('getWebViewUserAgent');
-  }
-
   static Future<String?> getAdServicesToken() {
     return _channel.invokeMethod<String>('getAdServicesToken');
-  }
-
-  static Future<bool> requestTrackingAuthorization() async {
-    return await _channel.invokeMethod<bool>('requestTrackingAuthorization') ??
-        false;
   }
 }